Move networking code into its own file and statically initialize the message queue
